Seção 9 - DTOs e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -4,7 +4,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
+using ProEventos.Application.Dtos;
+
 
 namespace ProEventos.API.Controllers
 {
@@ -23,8 +24,8 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado.");
-
+                if (eventos == null) return NoContent();
+                
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -39,9 +40,8 @@ namespace ProEventos.API.Controllers
         {
             try
             {
-                Evento evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Nenhum evento encontrado.");
-
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -55,8 +55,8 @@ namespace ProEventos.API.Controllers
         {
             try
             {
-                Evento[] eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (eventos == null) return NotFound("Eventos por tema não encontrados.");
+                var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -67,11 +67,11 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Evento model)
+        public async Task<IActionResult> Post([FromBody] EventoDto model)
         {
            try
             {
-                Evento evento = await _eventoService.AddEvento(model);
+                var evento = await _eventoService.AddEvento(model);
                 if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
 
                 return Ok();
@@ -83,14 +83,14 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
            try
             {
-                Evento evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao atualizar evento.");
+                EventoDto evento = await _eventoService.UpdateEvento(id, model);
+                if (evento == null) return NoContent();
 
-                return Ok();
+                return Ok(evento);
             }
             catch (Exception ex)
             {
@@ -102,6 +102,9 @@ namespace ProEventos.API.Controllers
         public async Task<IActionResult> Delete(int id){
            try
             {
+                var eventoDto = await _eventoService.GetEventoByIdAsync(id, true);
+                if (eventoDto == null) return NoContent();
+
                 Boolean evento = await _eventoService.DeleteEvento(id);
                 return evento ? Ok("Deletado") : BadRequest("Evento não deletado");
             }
